:sparkles: Make create button to save in server
