+++ packages/web/gql/gql.ts
@@ -14,6 +14,8 @@ import { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/
  */
 const documents = {
     "\n  query GetRecommendationForSubGoals(\n    $mainGoal: String!\n    $selectedSubGoals: [String!]\n    $currentLanguage: String!\n  ) {\n    recommendationForSubGoals(\n      mainGoal: $mainGoal\n      selectedSubGoals: $selectedSubGoals\n      currentLanguage: $currentLanguage\n    ) {\n      ... on RecommendationSuccess {\n        recommendations {\n          text\n        }\n      }\n      ... on RecommendationFailure {\n        errorType\n      }\n    }\n  }\n": types.GetRecommendationForSubGoalsDocument,
+    "\n  mutation CreateMandalaChart($input: CreateMandalaChartInput!) {\n    createMandalaChart(input: $input) {\n      ... on CreateMandalaChartSuccess {\n        mandalaChart {\n          _id\n        }\n      }\n      ... on CreateMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n": types.CreateMandalaChartDocument,
+    "\n  mutation UpdateMandalaChart($input: UpdateMandalaChartInput!) {\n    updateMandalaChart(input: $input) {\n      ... on UpdateMandalaChartSuccess {\n        mandalaChart {\n          _id\n        }\n      }\n      ... on UpdateMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n": types.UpdateMandalaChartDocument,
     "\n  mutation DeleteMandalaChart($input: DeleteMandalaChartInput!) {\n    deleteMandalaChart(input: $input) {\n      ... on DeleteMandalaChartSuccess {\n        _id\n      }\n      ... on DeleteMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n": types.DeleteMandalaChartDocument,
     "\n  mutation SignOut {\n    signOut\n  }\n": types.SignOutDocument,
     "\n  mutation DeleteAccount($email: String!) {\n    deleteAccount(email: $email) {\n      ... on DeleteAccountSuccess {\n        success\n      }\n      ... on DeleteAccountFailure {\n        errorType\n      }\n    }\n  }\n": types.DeleteAccountDocument,
@@ -23,8 +25,6 @@ const documents = {
     "\n  mutation SignUp(\n    $email: String!\n    $password: String!\n    $passwordConfirm: String!\n    $nickname: String!\n  ) {\n    signUp(\n      email: $email\n      password: $password\n      passwordConfirm: $passwordConfirm\n      nickname: $nickname\n    ) {\n      ... on SignUpSuccess {\n        isMailSent\n      }\n      ... on SignUpFailure {\n        errorType\n      }\n    }\n  }\n": types.SignUpDocument,
     "\n  mutation VerifyEmail($token: String!) {\n    verifyEmail(token: $token) {\n      ... on VerifyEmailSuccess {\n        success\n      }\n      ... on VerifyEmailFailure {\n        errorType\n      }\n    }\n  }\n": types.VerifyEmailDocument,
     "\n  query GetMandalaChart($input: GetMandalaChartInput!) {\n    getMandalaChart(input: $input) {\n      ... on GetMandalaChartSuccess {\n        mandalaChart {\n          _id\n          title\n          description\n          private\n          createdAt\n          lastModifiedAt\n          centerCell {\n            goal\n            tasks\n          }\n          surroundingCells {\n            goal\n            tasks\n          }\n        }\n      }\n      ... on GetMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n": types.GetMandalaChartDocument,
-    "\n  mutation CreateMandalaChart($input: CreateMandalaChartInput!) {\n    createMandalaChart(input: $input) {\n      ... on CreateMandalaChartSuccess {\n        mandalaChart {\n          _id\n        }\n      }\n      ... on CreateMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n": types.CreateMandalaChartDocument,
-    "\n  mutation UpdateMandalaChart($input: UpdateMandalaChartInput!) {\n    updateMandalaChart(input: $input) {\n      ... on UpdateMandalaChartSuccess {\n        mandalaChart {\n          _id\n        }\n      }\n      ... on UpdateMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n": types.UpdateMandalaChartDocument,
     "\n  query GetUserMandalaCharts($input: GetUserMandalaChartsInput!) {\n    getUserMandalaCharts(input: $input) {\n      ... on GetUserMandalaChartsSuccess {\n        mandalaCharts {\n          _id\n          title\n          description\n          private\n          createdAt\n          lastModifiedAt\n        }\n      }\n      ... on GetUserMandalaChartsFailure {\n        errorType\n      }\n    }\n  }\n": types.GetUserMandalaChartsDocument,
 };
 
@@ -46,6 +46,14 @@ export function graphql(source: string): unknown;
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  query GetRecommendationForSubGoals(\n    $mainGoal: String!\n    $selectedSubGoals: [String!]\n    $currentLanguage: String!\n  ) {\n    recommendationForSubGoals(\n      mainGoal: $mainGoal\n      selectedSubGoals: $selectedSubGoals\n      currentLanguage: $currentLanguage\n    ) {\n      ... on RecommendationSuccess {\n        recommendations {\n          text\n        }\n      }\n      ... on RecommendationFailure {\n        errorType\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetRecommendationForSubGoals(\n    $mainGoal: String!\n    $selectedSubGoals: [String!]\n    $currentLanguage: String!\n  ) {\n    recommendationForSubGoals(\n      mainGoal: $mainGoal\n      selectedSubGoals: $selectedSubGoals\n      currentLanguage: $currentLanguage\n    ) {\n      ... on RecommendationSuccess {\n        recommendations {\n          text\n        }\n      }\n      ... on RecommendationFailure {\n        errorType\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  mutation CreateMandalaChart($input: CreateMandalaChartInput!) {\n    createMandalaChart(input: $input) {\n      ... on CreateMandalaChartSuccess {\n        mandalaChart {\n          _id\n        }\n      }\n      ... on CreateMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n"): (typeof documents)["\n  mutation CreateMandalaChart($input: CreateMandalaChartInput!) {\n    createMandalaChart(input: $input) {\n      ... on CreateMandalaChartSuccess {\n        mandalaChart {\n          _id\n        }\n      }\n      ... on CreateMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n"];
+/**
+ * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
+ */
+export function graphql(source: "\n  mutation UpdateMandalaChart($input: UpdateMandalaChartInput!) {\n    updateMandalaChart(input: $input) {\n      ... on UpdateMandalaChartSuccess {\n        mandalaChart {\n          _id\n        }\n      }\n      ... on UpdateMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n"): (typeof documents)["\n  mutation UpdateMandalaChart($input: UpdateMandalaChartInput!) {\n    updateMandalaChart(input: $input) {\n      ... on UpdateMandalaChartSuccess {\n        mandalaChart {\n          _id\n        }\n      }\n      ... on UpdateMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
@@ -82,14 +90,6 @@ export function graphql(source: "\n  mutation VerifyEmail($token: String!) {\n  
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n  query GetMandalaChart($input: GetMandalaChartInput!) {\n    getMandalaChart(input: $input) {\n      ... on GetMandalaChartSuccess {\n        mandalaChart {\n          _id\n          title\n          description\n          private\n          createdAt\n          lastModifiedAt\n          centerCell {\n            goal\n            tasks\n          }\n          surroundingCells {\n            goal\n            tasks\n          }\n        }\n      }\n      ... on GetMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetMandalaChart($input: GetMandalaChartInput!) {\n    getMandalaChart(input: $input) {\n      ... on GetMandalaChartSuccess {\n        mandalaChart {\n          _id\n          title\n          description\n          private\n          createdAt\n          lastModifiedAt\n          centerCell {\n            goal\n            tasks\n          }\n          surroundingCells {\n            goal\n            tasks\n          }\n        }\n      }\n      ... on GetMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation CreateMandalaChart($input: CreateMandalaChartInput!) {\n    createMandalaChart(input: $input) {\n      ... on CreateMandalaChartSuccess {\n        mandalaChart {\n          _id\n        }\n      }\n      ... on CreateMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n"): (typeof documents)["\n  mutation CreateMandalaChart($input: CreateMandalaChartInput!) {\n    createMandalaChart(input: $input) {\n      ... on CreateMandalaChartSuccess {\n        mandalaChart {\n          _id\n        }\n      }\n      ... on CreateMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation UpdateMandalaChart($input: UpdateMandalaChartInput!) {\n    updateMandalaChart(input: $input) {\n      ... on UpdateMandalaChartSuccess {\n        mandalaChart {\n          _id\n        }\n      }\n      ... on UpdateMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n"): (typeof documents)["\n  mutation UpdateMandalaChart($input: UpdateMandalaChartInput!) {\n    updateMandalaChart(input: $input) {\n      ... on UpdateMandalaChartSuccess {\n        mandalaChart {\n          _id\n        }\n      }\n      ... on UpdateMandalaChartFailure {\n        errorType\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
